fix(analytics): 🐛 cast score to numeric before ROUND in contest queries
Postgres' two-arg ROUND only accepts numeric, not double precision.
The contest-analytics first-AC query and the leaderboard live-rank
+ summary CTEs all called ROUND(s.score, 2) directly on a double-
precision column, which raised "function round(double precision,
integer) does not exist" on environments that no longer auto-cast.
Cast through ::numeric to match Postgres 18 behaviour and to keep
the queries portable.

+++ src/lib/assignments/contest-analytics.ts
@@ -182,7 +182,7 @@ export async function computeContestAnalytics(assignmentId: string, includeTimel
     rawQueryAll<{ userId: string; problemId: string; submittedAt: Date }>(
       `SELECT s.user_id AS "userId", s.problem_id AS "problemId", s.submitted_at AS "submittedAt"
        FROM submissions s
-       WHERE s.assignment_id = @assignmentId AND ROUND(s.score, 2) = 100
+       WHERE s.assignment_id = @assignmentId AND ROUND(s.score::numeric, 2) = 100
        ORDER BY s.submitted_at ASC`,
       { assignmentId }
     ),
